+++ android/src/main/java/map/naver/plugin/net/lbstech/naver_map_plugin/NaverMarkerController.java
@@ -3,6 +3,8 @@ package map.naver.plugin.net.lbstech.naver_map_plugin;
 import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
+import android.util.DisplayMetrics;
+import android.view.WindowManager;
 
 import androidx.annotation.NonNull;
 
@@ -95,6 +97,12 @@ class NaverMarkerController {
         }
 
         void interpret(HashMap<String, Object> json) {
+            DisplayMetrics metrics = new DisplayMetrics();
+            WindowManager mgr = (WindowManager)context.getSystemService(Context.WINDOW_SERVICE);
+            mgr.getDefaultDisplay().getMetrics(metrics);
+
+            final float density = metrics.density;
+
             final Object position = json.get("position");
             if (position != null) marker.setPosition(Convert.toLatLng(position));
 
@@ -106,9 +114,6 @@ class NaverMarkerController {
 
             final Object anchor = json.get("anchor");
             if (anchor != null) marker.setAnchor(Convert.toPoint(anchor));
-
-            final Object isHideCollidedCaptions = json.get("isHideCollidedCaptions");
-            if (isHideCollidedCaptions != null) marker.setHideCollidedCaptions((Boolean) isHideCollidedCaptions);
 
             final Object captionText = json.get("captionText");
             if (captionText != null) marker.setCaptionText((String) captionText);
@@ -123,10 +128,10 @@ class NaverMarkerController {
             if (captionHaloColor != null) marker.setCaptionHaloColor(Convert.toColorInt(captionHaloColor));
 
             final Object width = json.get("width");
-            if (width != null) marker.setWidth(Math.round((int)width * density));
+            if (width != null) marker.setWidth(Math.round((int)width * NaverMarkerController.this.density));
 
             final Object height = json.get("height");
-            if (height != null) marker.setHeight(Math.round((int)height * density));
+            if (height != null) marker.setHeight(Math.round((int)height * NaverMarkerController.this.density));
 
             final Object maxZoom = json.get("maxZoom");
             if (maxZoom != null) marker.setMaxZoom((double)maxZoom);
@@ -147,7 +152,11 @@ class NaverMarkerController {
             if (captionMinZoom != null) marker.setCaptionMinZoom((double) captionMinZoom);
 
             final Object captionOffset = json.get("captionOffset");
-            if (captionOffset != null) marker.setCaptionOffset((int) captionOffset);
+            if (captionOffset != null) marker.setCaptionOffset(Math.round((int) captionOffset * density));
+
+            final Object isHideCollidedCaptions = json.get("isHideCollidedCaptions");
+            if (isHideCollidedCaptions != null)
+                marker.setHideCollidedCaptions((boolean) isHideCollidedCaptions);
 
             final Object captionPerspectiveEnabled = json.get("captionPerspectiveEnabled");
             if (captionPerspectiveEnabled != null)
@@ -177,7 +186,7 @@ class NaverMarkerController {
 
             final Object subCaptionRequestedWidth = json.get("subCaptionRequestedWidth");
             if (subCaptionRequestedWidth != null)
-                marker.setSubCaptionRequestedWidth(Math.round((int)subCaptionRequestedWidth * density));
+                marker.setSubCaptionRequestedWidth(Math.round((int)subCaptionRequestedWidth * NaverMarkerController.this.density));
 
             final Object icon = json.get("icon");
             if (icon != null) marker.setIcon(Convert.toOverlayImage(icon));