+++ lib/src/overlay_image.dart
@@ -11,9 +11,7 @@ class OverlayImage {
 
   /// ## [assetName] 이미지 중 [configuration]에 맞는 이미지를 찾아 [OverlayImage]객체를 만든다.
   ///
-  /// [context]를 통해 [ImageConfiguration]을 만들어 이미지를 찾는다.
   /// 이 때 [ImageConfiguration.bundle]은 [PlatformAssetBundle]이 된다.
-  /// [context]가 null일 경우 [ImageConfiguration.locale]과
   /// [ImageConfiguration.textDirection]은 null이 된다.
   ///
   /// - .../image.png  ---> android mdpi(1.0x)와 ios @1x 에서의 기본값
@@ -23,14 +21,21 @@ class OverlayImage {
   /// - .../4.0x/image.png  ---> android xxxhdpi(4.0x)에서 기본
   static Future<OverlayImage> fromAssetImage({
     required String assetName,
-    required BuildContext context,
+    AssetBundle? bundle,
+    double? devicePixelRatio,
+    Locale? locale,
+    TextDirection? textDirection,
+    Size? size,
+    TargetPlatform? platform,
   }) async {
     final _configuration = ImageConfiguration(
-      devicePixelRatio: MediaQuery.of(context).devicePixelRatio,
-      locale: context == null ? null : Localizations.localeOf(context),
-      textDirection: context == null ? null : Directionality.of(context),
-      platform: Platform.isIOS ? TargetPlatform.iOS : TargetPlatform.android,
-      bundle: PlatformAssetBundle(),
+      bundle: bundle,
+      devicePixelRatio: devicePixelRatio,
+      locale: locale,
+      textDirection: textDirection,
+      size: size,
+      platform: platform ??
+          (Platform.isIOS ? TargetPlatform.iOS : TargetPlatform.android),
     );
     final AssetImage assetImage = AssetImage(assetName);
     final AssetBundleImageKey key = await assetImage.obtainKey(_configuration);